backend/requirements.txt
blinker==1.6.2
click==8.1.3
flake8==6.0.0
Flask==2.1.0
gunicorn==20.0.4
importlib-metadata==6.6.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
nulltype==2.3.1
plaid-python==9.1.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.15.0
PyYAML==6.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
